refactor: move commands to cpp
