+++ src/server/src/extensions/wm/switch-windows-command.hpp
@@ -3,13 +3,4 @@
 #include "single-view-command-context.hpp"
 #include "theme.hpp"
 #include <QCoreApplication>
-
-class SwitchWindowsCommand : public BuiltinViewCommand<SwitchWindowsViewHost> {
-  QString id() const override { return "switch-windows"; }
-  QString name() const override {
-    return QCoreApplication::translate("SwitchWindowsCommand", "Switch Windows");
-  }
-  ImageURL iconUrl() const override {
-    return ImageURL::builtin(BuiltinIcon::AppWindowList).setBackgroundTint(SemanticColor::Cyan);
-  }
-};
+;
